refact: [community/profile-list] modifie le design de la page
