+++ backend-implicaction/src/main/java/com/dynonuggets/refonteimplicaction/community/profile/controller/ProfileController.java
@@ -5,6 +5,9 @@ import com.dynonuggets.refonteimplicaction.community.profile.dto.ProfileUpdateRe
 import com.dynonuggets.refonteimplicaction.community.profile.service.ProfileService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -14,7 +17,6 @@ import javax.validation.Valid;
 import static com.dynonuggets.refonteimplicaction.community.profile.utils.ProfileUris.GET_PROFILE_BY_USERNAME;
 import static com.dynonuggets.refonteimplicaction.community.profile.utils.ProfileUris.PROFILES_BASE_URI;
 import static com.dynonuggets.refonteimplicaction.filemanagement.utils.FileUris.POST_PROFILE_AVATAR;
-import static org.springframework.data.domain.PageRequest.of;
 
 @RestController
 @AllArgsConstructor
@@ -26,9 +28,12 @@ public class ProfileController {
     @GetMapping
     public ResponseEntity<Page<ProfileDto>> getAllProfiles(
             @RequestParam(defaultValue = "0") final int page,
-            @RequestParam(defaultValue = "10") final int rows
+            @RequestParam(defaultValue = "10") final int rows,
+            @RequestParam(value = "sortBy", defaultValue = "id") final String[] sortBy,
+            @RequestParam(value = "sortOrder", defaultValue = "ASC") final String sortOrder
     ) {
-        return ResponseEntity.ok(profileService.getAllProfiles(of(page, rows)));
+        final Pageable pageable = PageRequest.of(page, rows, Sort.by(Sort.Direction.valueOf(sortOrder), sortBy));
+        return ResponseEntity.ok(profileService.getAllProfiles(pageable));
     }
 
     @PutMapping
