+++ frontend-implicaction/src/app/community/components/profile-details/tabs/edit-profile-tab/edit-profile-tab.component.ts
@@ -1,10 +1,10 @@
 import {Component, Input, OnDestroy} from '@angular/core';
-import {Profile} from '../../../../models/profile/profile';
+import {Profile} from '../../../../models/profile';
 import {ProfileContextService} from '../../../../../core/services/profile-context.service';
 import {ProfileService} from '../../../../services/profile/profile.service';
 import {ToasterService} from '../../../../../core/services/toaster.service';
 import {Subject} from 'rxjs';
-import {ProfileUpdateRequest} from '../../../../models/profile/profile-update-request';
+import {ProfileUpdateRequest} from '../../../../models/profile-update-request';
 import {finalize} from 'rxjs/operators';
 
 @Component({
