+++ frontend-implicaction/src/app/community/components/profile-details/tabs/profile-overview/profile-details-section/profile-details-section.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input} from '@angular/core';
-import {Profile} from "../../../../../models/profile/profile";
+import {Profile} from "../../../../../models/profile";
 
 @Component({
   selector: 'app-profile-details-section',
