+++ frontend-implicaction/src/app/community/components/profile-details/tabs/profile-overview/profile-overview-tab/profile-overview-tab.component.ts
@@ -1,5 +1,5 @@
 import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
-import {Profile} from "../../../../../models/profile/profile";
+import {Profile} from "../../../../../models/profile";
 
 @Component({
   selector: 'app-profile-overview-tab',
