+++ frontend-implicaction/src/app/community/models/enums/relation-action-button-enum.ts
@@ -1,4 +1,4 @@
-import {EnumCodeLabelAbstract} from "../../../../shared/enums/enum-code-label-abstract.enum";
+import {EnumCodeLabelAbstract} from "../../../shared/enums/enum-code-label-abstract.enum";
 
 export enum RelationActionButton {
   RELATION_DELETE = 'RELATION_CANCEL',
