+++ frontend-implicaction/src/app/community/models/relation-action.ts
@@ -1,5 +1,5 @@
 import {Relation} from "./relation";
-import {EnumCodeLabelAbstract} from "../../../shared/enums/enum-code-label-abstract.enum";
+import {EnumCodeLabelAbstract} from "../../shared/enums/enum-code-label-abstract.enum";
 
 export enum RelationActionEnumCode {
   CREATE = 'CREATE', DELETE = 'DELETE', CONFIRM = 'CONFIRM'
