+++ frontend-implicaction/src/app/community/models/relation.ts
@@ -1,0 +1,9 @@
+import {Profile} from "./profile";
+
+export interface Relation {
+  id?: string;
+  confirmedAt?: string;
+  sentAt?: string;
+  receiver?: Profile;
+  sender?: Profile;
+}
