--- frontend-implicaction/src/app/community/models/relation/relation-type.enum.ts
@@ -1,6 +1,0 @@
-export enum RelationTypeCode {
-  FRIEND = 'FRIEND',
-  RECEIVER = 'RECEIVER',
-  SENDER = 'SENDER',
-  NONE = 'NONE'
-}
