+++ frontend-implicaction/src/app/community/services/profile/profile.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from '@angular/core';
 import {Observable} from "rxjs";
-import {ProfileUpdateRequest} from "../../models/profile/profile-update-request";
+import {ProfileUpdateRequest} from "../../models/profile-update-request";
 import {HttpClient, HttpEvent, HttpRequest} from "@angular/common/http";
 import {ApiEndpointsService} from "../../../core/services/api-endpoints.service";
-import {Profile} from "../../models/profile/profile";
+import {Profile} from "../../models/profile";
 import {Pageable} from "../../../shared/models/pageable";
 
 @Injectable({
